Initial commit - Complete app

+++ app/src/main/java/com/example/electricitybillcalculator2/DatabaseHelper.java
@@ -12,10 +12,9 @@ import java.util.List;
 public class DatabaseHelper extends SQLiteOpenHelper {
 
     private static final String DATABASE_NAME = "ElectricityBill.db";
-    private static final int DATABASE_VERSION = 1;
+    private static final int DATABASE_VERSION = 2; // Changed to version 2
     private static final String TABLE_NAME = "bills";
 
-    // Kolom tabel
     private static final String COLUMN_ID = "id";
     private static final String COLUMN_MONTH = "month";
     private static final String COLUMN_UNIT = "unit";
@@ -24,13 +23,12 @@ public class DatabaseHelper extends SQLiteOpenHelper {
     private static final String COLUMN_FINAL_COST = "final_cost";
     private static final String COLUMN_CREATED_AT = "created_at";
 
-    // Query pembuatan tabel
     private static final String CREATE_TABLE = "CREATE TABLE " + TABLE_NAME + "("
             + COLUMN_ID + " INTEGER PRIMARY KEY AUTOINCREMENT,"
             + COLUMN_MONTH + " TEXT,"
             + COLUMN_UNIT + " REAL,"
             + COLUMN_TOTAL_CHARGES + " REAL,"
-            + COLUMN_REBATE + " INTEGER,"
+            + COLUMN_REBATE + " REAL," // Changed from INTEGER to REAL
             + COLUMN_FINAL_COST + " REAL,"
             + COLUMN_CREATED_AT + " DATETIME DEFAULT CURRENT_TIMESTAMP"
             + ")";
@@ -46,11 +44,14 @@ public class DatabaseHelper extends SQLiteOpenHelper {
 
     @Override
     public void onUpgrade(SQLiteDatabase db, int oldVersion, int newVersion) {
-        db.execSQL("DROP TABLE IF EXISTS " + TABLE_NAME);
-        onCreate(db);
+        if (oldVersion < 2) {
+            // If upgrading from version 1 to 2, we need to recreate the table
+            // with REAL type for rebate column
+            db.execSQL("DROP TABLE IF EXISTS " + TABLE_NAME);
+            onCreate(db);
+        }
     }
 
-    // Tambah data bill
     public long addBill(BillModel bill) {
         SQLiteDatabase db = this.getWritableDatabase();
         ContentValues values = new ContentValues();
@@ -66,7 +67,6 @@ public class DatabaseHelper extends SQLiteOpenHelper {
         return id;
     }
 
-    // Ambil semua bill
     public List<BillModel> getAllBills() {
         List<BillModel> billList = new ArrayList<>();
         String selectQuery = "SELECT * FROM " + TABLE_NAME + " ORDER BY " + COLUMN_CREATED_AT + " DESC";
@@ -81,7 +81,7 @@ public class DatabaseHelper extends SQLiteOpenHelper {
                 bill.setMonth(cursor.getString(cursor.getColumnIndexOrThrow(COLUMN_MONTH)));
                 bill.setUnit(cursor.getDouble(cursor.getColumnIndexOrThrow(COLUMN_UNIT)));
                 bill.setTotalCharges(cursor.getDouble(cursor.getColumnIndexOrThrow(COLUMN_TOTAL_CHARGES)));
-                bill.setRebate(cursor.getInt(cursor.getColumnIndexOrThrow(COLUMN_REBATE)));
+                bill.setRebate(cursor.getDouble(cursor.getColumnIndexOrThrow(COLUMN_REBATE))); // Changed to getDouble
                 bill.setFinalCost(cursor.getDouble(cursor.getColumnIndexOrThrow(COLUMN_FINAL_COST)));
 
                 billList.add(bill);
@@ -93,10 +93,9 @@ public class DatabaseHelper extends SQLiteOpenHelper {
         return billList;
     }
 
-    // Ambil bill berdasarkan ID
     public BillModel getBill(int id) {
         SQLiteDatabase db = this.getReadableDatabase();
-        BillModel bill = new BillModel();
+        BillModel bill = null; // Changed to null initially
 
         Cursor cursor = db.query(TABLE_NAME,
                 new String[]{COLUMN_ID, COLUMN_MONTH, COLUMN_UNIT, COLUMN_TOTAL_CHARGES, COLUMN_REBATE, COLUMN_FINAL_COST},
@@ -105,11 +104,12 @@ public class DatabaseHelper extends SQLiteOpenHelper {
                 null, null, null, null);
 
         if (cursor != null && cursor.moveToFirst()) {
+            bill = new BillModel(); // Initialize here
             bill.setId(cursor.getInt(cursor.getColumnIndexOrThrow(COLUMN_ID)));
             bill.setMonth(cursor.getString(cursor.getColumnIndexOrThrow(COLUMN_MONTH)));
             bill.setUnit(cursor.getDouble(cursor.getColumnIndexOrThrow(COLUMN_UNIT)));
             bill.setTotalCharges(cursor.getDouble(cursor.getColumnIndexOrThrow(COLUMN_TOTAL_CHARGES)));
-            bill.setRebate(cursor.getInt(cursor.getColumnIndexOrThrow(COLUMN_REBATE)));
+            bill.setRebate(cursor.getDouble(cursor.getColumnIndexOrThrow(COLUMN_REBATE))); // Changed to getDouble
             bill.setFinalCost(cursor.getDouble(cursor.getColumnIndexOrThrow(COLUMN_FINAL_COST)));
             cursor.close();
         }
@@ -118,10 +118,40 @@ public class DatabaseHelper extends SQLiteOpenHelper {
         return bill;
     }
 
-    // Hapus semua data
     public void deleteAllBills() {
         SQLiteDatabase db = this.getWritableDatabase();
         db.delete(TABLE_NAME, null, null);
         db.close();
+    }
+
+    // ADD THESE NEW METHODS
+
+    public boolean updateBill(BillModel bill) {
+        SQLiteDatabase db = this.getWritableDatabase();
+        ContentValues values = new ContentValues();
+
+        values.put(COLUMN_MONTH, bill.getMonth());
+        values.put(COLUMN_UNIT, bill.getUnit());
+        values.put(COLUMN_TOTAL_CHARGES, bill.getTotalCharges());
+        values.put(COLUMN_REBATE, bill.getRebate());
+        values.put(COLUMN_FINAL_COST, bill.getFinalCost());
+
+        int rowsAffected = db.update(TABLE_NAME,
+                values,
+                COLUMN_ID + " = ?",
+                new String[]{String.valueOf(bill.getId())});
+
+        db.close();
+        return rowsAffected > 0;
+    }
+
+    public boolean deleteBill(int id) {
+        SQLiteDatabase db = this.getWritableDatabase();
+        int rowsAffected = db.delete(TABLE_NAME,
+                COLUMN_ID + " = ?",
+                new String[]{String.valueOf(id)});
+
+        db.close();
+        return rowsAffected > 0;
     }
 }
